pyproject.toml
[tool.poetry]
name = "cyclonedx"
version = "0.1.0"
description = "Api to ingest and process SBOMS"
authors = ["jon smith <jonsmith@example.com>"]

[tool.poetry.scripts]
clean = "deploy.build:clean"
build = "deploy.build:run"
lint = "deploy.build:lint"
test = "deploy.build:test"
package = "deploy.build:package"
deploy = "deploy.deploy:run"
tokentest = "tests.e2e:token_test"
teamtest = "tests.e2e:team_test"
sbomuploadtest = "tests.e2e:sbom_upload_test"
usersearchtest = "tests.e2e:user_search_test"
dodep = "deploy.deploy:dodep"
setupadmin = "deploy.deploy:setup_admin_user"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "^1.21.10"
jsonschema = "^4.4.0"
requests = "^2.27.1"
requests-toolbelt = "^0.9.1"
urllib3 = "^1.26.9"
"aws-cdk.aws-apigatewayv2-authorizers-alpha" = "^2.22.0-alpha.0"
"aws-cdk.aws-apigatewayv2-integrations-alpha" = "^2.22.0-alpha.0"
genson = "^1.2.2"
python-dotenv = "^0.20.0"
dparse = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-mock = "^3.7.0"
black = "^22.1.0"
pre-commit = "^2.17.0"
pylint = "^2.12.2"
pytest-cov = "^3.0.0"
aws-cdk-lib = "^2.17.0"
ipython = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
